rename: 이동 notification이라 잘못 선언되어 있던 패키지 review로
Related to: #69
