+++ be/review-api/src/main/java/com/zzaug/review/web/support/usecase/ReviewUseCaseRequestConverter.java
@@ -1,7 +1,7 @@
 package com.zzaug.review.web.support.usecase;
 
 import com.zzaug.review.domain.dto.review.ReviewUseCaseRequest;
-import com.zzaug.review.web.dto.notification.ReviewRequest;
+import com.zzaug.review.web.dto.review.ReviewRequest;
 import lombok.experimental.UtilityClass;
 
 @UtilityClass
